+++ Source/DBProfileViewControllerDefaults.m
@@ -13,7 +13,8 @@
 @implementation DBProfileViewControllerDefaults
 
 + (UIColor *)defaultSegmentedControlTintColor {
-    return [UIColor grayColor];
+    //return [UIColor grayColor];
+    return [UIColor colorWithRed:29/255.0 green:161/255.0 blue:242/255.0 alpha:1];
 }
 
 + (UIImage *)defaultBackBarButtonItemImageForTraitCollection:(UITraitCollection *)traitCollection {