+++ Source/Views/Private/DBProfileTintedImageView.h
@@ -1,9 +1,11 @@
 //
 //  DBProfileTintedImageView.h
-//  Pods
+//  DBProfileViewController
 //
 //  Created by Devon Boyer on 2016-03-21.
+//  Copyright (c) 2015 Devon Boyer. All rights reserved.
 //
+//  Released under an MIT license: http://opensource.org/licenses/MIT
 //
 
 #import <UIKit/UIKit.h>